requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.1
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
iopath==0.1.10
ipykernel==6.25.2
ipython
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.299
langsmith==0.0.40
layoutparser==0.3.4
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.6
numpy
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
portalocker==2.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.1.1
pypdfium2
pytesseract==0.3.10
python-dateutil==2.8.2
python-iso639==2023.6.15
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rapidfuzz==3.3.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.10.16
unstructured-inference==0.6.1
uri-template==1.3.0
urllib3==2.0.5
userpaths==0.1.3
validators==0.22.0
wcwidth==0.2.6
weaviate-client==3.24.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
xformer==1.0.1
yarl==1.9.2
